renames models/combined to models/apiReponse
